NavA3/requirements.txt
absl-py==2.2.2
accelerate==1.7.0
actionlib==1.14.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-cors==0.8.1
aiosignal==1.3.2
airportsdata==20250224
angles==1.9.13
annotated-types==0.7.0
anyio==4.9.0
astor==0.8.1
async-timeout==5.0.1
attrs==25.3.0
av==14.3.0
blake3==1.0.4
blinker==1.9.0
bondpy==1.8.6
cachetools==5.5.2
camera-calibration==1.17.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
cloudpickle==3.1.1
colorful==0.5.6
compressed-tensors==0.9.1
contourpy==1.3.2
controller-manager==0.20.0
controller-manager-msgs==0.20.0
cv-bridge==1.16.2
cycler==0.12.1
deepspeed==0.16.7
depyf==0.18.0
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
dill==0.4.0
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
docker-pycreds==0.4.0
dynamic-reconfigure==1.7.3
einops==0.8.1
exceptiongroup==1.3.0
fastapi==0.115.12
filelock==3.18.0
flash-attn @ file:///home/alan/%E4%B8%8B%E8%BD%BD/flash_attn-2.7.3%2Bcu11torch2.5cxx11abiFALSE-cp310-cp310-linux_x86_64.whl#sha256=90b92074ec1ae12b96c0ce4c8c3c9078b9a52422bcf3f4233e865d4bccee3edc
Flask==3.1.1
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.3.2
gazebo_plugins==2.9.2
gazebo_ros==2.9.2
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
gguf==0.10.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-auth==2.40.1
googleapis-common-protos==1.70.0
grpcio==1.71.0
h11==0.16.0
hjson==3.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.31.2
idna==3.10
image-geometry==1.16.2
importlib_metadata==8.7.0
interactive-markers==1.12.0
interegular==0.3.3
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.0
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lark==1.2.2
laser_geometry==1.6.7
lm-format-enforcer==0.10.11
MarkupSafe==3.0.2
matplotlib==3.10.3
message-filters==1.17.0
mistral_common==1.5.4
moveit-commander==1.1.16
moveit-core==1.1.16
moveit-python==0.4.5
moveit-ros-planning-interface==1.1.16
moveit-ros-visualization==1.1.16
moveit-task-constructor-core==0.1.3
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.4.3
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.78.1
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python-headless==4.11.0.86
outlines==0.1.11
outlines_core==0.1.26
packaging==25.0
partial-json-parser==0.2.1.1.post5
pillow==11.2.1
platformdirs==4.3.8
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.21.1
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
py-cpuinfo==9.0.0
py-spy==0.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycountry==24.6.1
pydantic==2.11.4
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-qt-binding==0.4.4
PyYAML==6.0.2
pyzmq==26.4.0
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
qwen-vl-utils==0.0.11
ray==2.46.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
resource_retriever==1.12.7
rosbag==1.17.0
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.17.0
roslaunch==1.17.0
roslib==1.15.8
roslint==0.12.0
roslz4==1.17.0
rosmake==1.15.8
rosmaster==1.17.0
rosmsg==1.17.0
rosnode==1.17.0
rosparam==1.17.0
rospy==1.17.0
rosservice==1.17.0
rostest==1.17.0
rostopic==1.17.0
rosunit==1.15.8
roswtf==1.17.0
rouge_score==0.1.2
rpds-py==0.24.0
rqt-console==0.4.12
rqt-image-view==0.4.17
rqt-logger-level==0.4.12
rqt-moveit==0.5.11
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.15
rqt-runtime-monitor==0.5.10
rqt-rviz==0.7.0
rqt-tf-tree==0.6.4
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_dep==0.4.12
rqt_graph==0.4.14
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_launch==0.4.9
rqt_msg==0.4.10
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.11
rqt_publisher==0.4.10
rqt_py_common==0.5.3
rqt_py_console==0.4.10
rqt_robot_steering==0.5.12
rqt_service_caller==0.4.10
rqt_shell==0.4.11
rqt_srv==0.4.9
rqt_top==0.4.10
rqt_topic==0.4.13
rqt_web==0.4.10
rsa==4.9.1
rviz==1.14.25
safetensors==0.5.3
sensor-msgs==1.13.1
sentencepiece==0.2.0
sentry-sdk==2.28.0
setproctitle==1.3.6
six==1.17.0
smach==2.5.3
smach-ros==2.5.3
smart-open==7.1.0
smclib==1.8.6
smmap==5.0.2
sniffio==1.3.1
srdfdom==0.6.4
starlette==0.46.2
sympy==1.13.1
tensorboardX==2.6.2.2
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.7
tf2-kdl==0.7.7
tf2-py==0.7.7
tf2-ros==0.7.7
tiktoken==0.9.0
tokenizers==0.21.1
topic-tools==1.17.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.51.3
triton==3.1.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urdfdom-py==0.4.6
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
virtualenv==20.31.2
vllm==0.7.2
wandb==0.18.3
watchfiles==1.0.5
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
xacro==1.14.18
xformers==0.0.28.post3
xgrammar==0.1.19
yarl==1.20.0
zipp==3.21.0
